requirements.txt
pypdf==4.0.1
python-docx==0.8.11
openpyxl==3.10.10
requests==2.31.0
